+++ tests/test_wsireader.py
@@ -202,11 +202,11 @@ def command_line_slide_thumbnail(runner, sample, tmp_path, mode="save"):
         cli.main,
         [
             "slide-thumbnail",
-            "--img_input",
+            "--img-input",
             str(pathlib.Path(sample)),
             "--mode",
             mode,
-            "--output_path",
+            "--output-path",
             str(pathlib.Path(tmp_path).joinpath("slide_thumb.jpg")),
         ],
     )
@@ -1565,7 +1565,7 @@ def test_command_line_read_bounds(sample_ndpi, tmp_path):
         cli.main,
         [
             "read-bounds",
-            "--img_input",
+            "--img-input",
             str(pathlib.Path(sample_ndpi)),
             "--resolution",
             "0",
@@ -1578,7 +1578,7 @@ def test_command_line_read_bounds(sample_ndpi, tmp_path):
             "0",
             "2000",
             "2000",
-            "--output_path",
+            "--output-path",
             str(pathlib.Path(tmp_path).joinpath("im_region.jpg")),
         ],
     )
@@ -1590,7 +1590,7 @@ def test_command_line_read_bounds(sample_ndpi, tmp_path):
         cli.main,
         [
             "read-bounds",
-            "--img_input",
+            "--img-input",
             str(pathlib.Path(sample_ndpi)),
             "--resolution",
             "0",
@@ -1598,7 +1598,7 @@ def test_command_line_read_bounds(sample_ndpi, tmp_path):
             "level",
             "--mode",
             "save",
-            "--output_path",
+            "--output-path",
             str(pathlib.Path(tmp_path).joinpath("im_region2.jpg")),
         ],
     )
@@ -1614,7 +1614,7 @@ def test_command_line_jp2_read_bounds(sample_jp2, tmp_path):
         cli.main,
         [
             "read-bounds",
-            "--img_input",
+            "--img-input",
             str(pathlib.Path(sample_jp2)),
             "--resolution",
             "0",
@@ -1636,7 +1636,7 @@ def test_command_line_jp2_read_bounds_show(sample_jp2, tmp_path):
         cli.main,
         [
             "read-bounds",
-            "--img_input",
+            "--img-input",
             str(pathlib.Path(sample_jp2)),
             "--resolution",
             "0",
@@ -1657,7 +1657,7 @@ def test_command_line_unsupported_file_read_bounds(sample_svs, tmp_path):
         cli.main,
         [
             "read-bounds",
-            "--img_input",
+            "--img-input",
             str(pathlib.Path(sample_svs))[:-1],
             "--resolution",
             "0",
@@ -1687,7 +1687,7 @@ def test_command_line_slide_thumbnail_output_none(sample_svs, tmp_path):
         cli.main,
         [
             "slide-thumbnail",
-            "--img_input",
+            "--img-input",
             str(pathlib.Path(sample_svs)),
             "--mode",
             "save",
@@ -1722,11 +1722,11 @@ def test_command_line_jp2_slide_thumbnail_file_not_supported(sample_jp2, tmp_pat
         cli.main,
         [
             "slide-thumbnail",
-            "--img_input",
+            "--img-input",
             str(pathlib.Path(sample_jp2))[:-1],
             "--mode",
             "save",
-            "--output_path",
+            "--output-path",
             str(pathlib.Path(tmp_path).joinpath("slide_thumb.jpg")),
         ],
     )